+++ tests/test_client.py
@@ -2,6 +2,7 @@ import json
 import uuid
 from datetime import datetime
 
+import pkg_resources
 import pytest
 from asynctest import patch
 from google.oauth2 import service_account
@@ -16,7 +17,6 @@ from async_firebase.messages import (
     Aps,
     ApsAlert,
     Message,
-    MulticastMessage,
 )
 
 pytestmark = pytest.mark.asyncio
@@ -135,6 +135,8 @@ async def test__prepare_headers(fake_async_fcm_client_w_creds):
             "Authorization": "Bearer fake-jwt-token",
             "Content-Type": "application/json; UTF-8",
             "X-Request-Id": str(frozen_uuid),
+            "X-GOOG-API-FORMAT-VERSION": "2",
+            "X-FIREBASE-CLIENT": 'async-firebase/{0}'.format(pkg_resources.get_distribution('async-firebase').version),
         }
 
 
@@ -265,19 +267,31 @@ async def test_push_realistic_payload(fake_async_fcm_client_w_creds, fake_device
     }
 
 
-@pytest.mark.parametrize("fake_multi_device_tokens", (5,), indirect=True)
+@pytest.mark.parametrize("fake_multi_device_tokens", (3,), indirect=True)
 async def test_push_multicast(fake_async_fcm_client_w_creds, fake_multi_device_tokens, httpx_mock: HTTPXMock):
     fake_async_fcm_client_w_creds._get_access_token = fake__get_access_token
     creds = fake_async_fcm_client_w_creds._credentials
+    response_data = (
+        '\r\n--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-4440d691-7909-4346-af9a-b44f17638f6c\r\n\r\nHTTP/1.1 200 OK\r\nContent-Type: '
+        'application/json; charset=UTF-8\r\nVary: Origin\r\nVary: X-Origin\r\nVary: Referer\r\n\r\n{\n  '
+        f'"name": "projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb24"\n}}\n\r\n'
+        '--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-fdbc3fd2-4031-4c00-88d2-22c9523bb941\r\n\r\nHTTP/1.1 200 OK\r\nContent-Type: '
+        'application/json; charset=UTF-8\r\nVary: Origin\r\nVary: X-Origin\r\nVary: Referer\r\n\r\n{\n  '
+        f'"name": "projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb25"\n}}\n\r\n'
+        '--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-222748d1-1388-4c06-a48f-445f7aef19a9\r\n\r\nHTTP/1.1 200 OK\r\nContent-Type: '
+        'application/json; charset=UTF-8\r\nVary: Origin\r\nVary: X-Origin\r\nVary: Referer\r\n\r\n{\n  '
+        f'"name": "projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb26"\n}}\n\r\n'
+        '--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4--\r\n'
+    )
+
     httpx_mock.add_response(
         status_code=200,
-        json=[
-            {"name": f"projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb24"},
-            {"name": f"projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb25"},
-            {"name": f"projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb26"},
-            {"name": f"projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb27"},
-            {"name": f"projects/{creds.project_id}/messages/0:1612788010922733%7606eb247606eb28"},
-        ]
+        data=response_data.encode(),
+        headers={"content-type": "multipart/mixed; boundary=batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4"}
+
     )
     apns_config = fake_async_fcm_client_w_creds.build_apns_config(
         priority="normal",
@@ -294,24 +308,32 @@ async def test_push_multicast(fake_async_fcm_client_w_creds, fake_multi_device_t
         {"name": "projects/fake-mobile-app/messages/0:1612788010922733%7606eb247606eb24"},
         {"name": "projects/fake-mobile-app/messages/0:1612788010922733%7606eb247606eb25"},
         {"name": "projects/fake-mobile-app/messages/0:1612788010922733%7606eb247606eb26"},
-        {"name": "projects/fake-mobile-app/messages/0:1612788010922733%7606eb247606eb27"},
-        {"name": "projects/fake-mobile-app/messages/0:1612788010922733%7606eb247606eb28"},
     ]
 
 
-@pytest.mark.parametrize("fake_multi_device_tokens", (5,), indirect=True)
+@pytest.mark.parametrize("fake_multi_device_tokens", (3,), indirect=True)
 async def test_push_multicast_dry_run(fake_async_fcm_client_w_creds, fake_multi_device_tokens, httpx_mock: HTTPXMock):
     fake_async_fcm_client_w_creds._get_access_token = fake__get_access_token
     creds = fake_async_fcm_client_w_creds._credentials
+    response_data = (
+        '\r\n--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-4440d691-7909-4346-af9a-b44f17638f6c\r\n\r\nHTTP/1.1 200 OK\r\nContent-Type: '
+        'application/json; charset=UTF-8\r\nVary: Origin\r\nVary: X-Origin\r\nVary: Referer\r\n\r\n{\n  '
+        f'"name": "projects/{creds.project_id}/messages/fake_message_id"\n}}\n\r\n'
+        '--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-fdbc3fd2-4031-4c00-88d2-22c9523bb941\r\n\r\nHTTP/1.1 200 OK\r\nContent-Type: '
+        'application/json; charset=UTF-8\r\nVary: Origin\r\nVary: X-Origin\r\nVary: Referer\r\n\r\n{\n  '
+        f'"name": "projects/{creds.project_id}/messages/fake_message_id"\n}}\n\r\n'
+        '--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-222748d1-1388-4c06-a48f-445f7aef19a9\r\n\r\nHTTP/1.1 200 OK\r\nContent-Type: '
+        'application/json; charset=UTF-8\r\nVary: Origin\r\nVary: X-Origin\r\nVary: Referer\r\n\r\n{\n  '
+        f'"name": "projects/{creds.project_id}/messages/fake_message_id"\n}}\n\r\n'
+        '--batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4--\r\n'
+    )
     httpx_mock.add_response(
         status_code=200,
-        json=[
-            {"name": f"projects/{creds.project_id}/messages/fake_message_id_1"},
-            {"name": f"projects/{creds.project_id}/messages/fake_message_id_2"},
-            {"name": f"projects/{creds.project_id}/messages/fake_message_id_3"},
-            {"name": f"projects/{creds.project_id}/messages/fake_message_id_4"},
-            {"name": f"projects/{creds.project_id}/messages/fake_message_id_5"},
-        ]
+        data=response_data.encode(),
+        headers={"content-type": "multipart/mixed; boundary=batch_llG_9dniIyeFXPERplIRPwpVYtn3RBa4"}
     )
     apns_config = fake_async_fcm_client_w_creds.build_apns_config(
         priority="normal",
@@ -325,11 +347,9 @@ async def test_push_multicast_dry_run(fake_async_fcm_client_w_creds, fake_multi_
         device_tokens=fake_multi_device_tokens, apns=apns_config, dry_run=True
     )
     assert response == [
-        {"name": "projects/fake-mobile-app/messages/fake_message_id_1"},
-        {"name": "projects/fake-mobile-app/messages/fake_message_id_2"},
-        {"name": "projects/fake-mobile-app/messages/fake_message_id_3"},
-        {"name": "projects/fake-mobile-app/messages/fake_message_id_4"},
-        {"name": "projects/fake-mobile-app/messages/fake_message_id_5"},
+        {"name": "projects/fake-mobile-app/messages/fake_message_id"},
+        {"name": "projects/fake-mobile-app/messages/fake_message_id"},
+        {"name": "projects/fake-mobile-app/messages/fake_message_id"},
     ]
 
 
@@ -353,20 +373,21 @@ async def test_push_multicast_too_many_tokens(
         )
 
 
-async def test_push_multicast_unauthenticated(fake_async_fcm_client_w_creds, httpx_mock: HTTPXMock):
+async def test_push_multicast_unknown_registration_token(fake_async_fcm_client_w_creds, httpx_mock: HTTPXMock):
     fake_async_fcm_client_w_creds._get_access_token = fake__get_access_token
+    response_data = (
+        '\r\n--batch_HwFDZe-SUCq5qEgCavJPhhi8tA7xJBlB\r\nContent-Type: application/http\r\nContent-ID: '
+        'response-363ad2c9-a3d1-45f5-b559-6d69a13a880e\r\n\r\nHTTP/1.1 400 Bad Request\r\nVary: Origin\r\nVary: '
+        'X-Origin\r\nVary: Referer\r\nContent-Type: application/json; charset=UTF-8\r\n\r\n{\n  "error": {\n    '
+        '"code": 400,\n    "message": "The registration token is not a valid FCM registration token",\n    "status": '
+        '"INVALID_ARGUMENT",\n    "details": [\n      {\n        "@type": '
+        '"type.googleapis.com/google.firebase.fcm.v1.FcmError",\n        "errorCode": "INVALID_ARGUMENT"\n      }\n   '
+        ' ]\n  }\n}\n\r\n--batch_HwFDZe-SUCq5qEgCavJPhhi8tA7xJBlB--\r\n'
+    )
     httpx_mock.add_response(
-        status_code=401,
-        json={
-            "error": {
-                "code": 401,
-                "message": "Request had invalid authentication credentials. "
-                "Expected OAuth 2 access token, login cookie or other "
-                "valid authentication credential. See "
-                "https://developers.google.com/identity/sign-in/web/devconsole-project.",
-                "status": "UNAUTHENTICATED",
-            }
-        },
+        status_code=400,
+        data=response_data.encode(),
+        headers={"content-type": "multipart/mixed; boundary=batch_HwFDZe-SUCq5qEgCavJPhhi8tA7xJBlB"}
     )
     apns_config = fake_async_fcm_client_w_creds.build_apns_config(
         priority="normal",
@@ -377,7 +398,7 @@ async def test_push_multicast_unauthenticated(fake_async_fcm_client_w_creds, htt
         custom_data={"foo": "bar"},
     )
     response = await fake_async_fcm_client_w_creds.push_multicast(device_tokens=["qwerty:ytrewq"], apns=apns_config)
-    assert response["error"]["code"] == 401
+    assert response[0]["error"]["code"] == 400
 
 
 async def test_push_multicast_data_has_not_provided(fake_async_fcm_client_w_creds):
@@ -432,67 +453,6 @@ async def test_push_multicast_data_has_not_provided(fake_async_fcm_client_w_cred
             {
                 "message": {
                     "token": "token-1",
-                    "apns": {
-                        "payload": {
-                            "aps": {
-                                "alert": "push-text",
-                                "badge": 5,
-                                "sound": "default",
-                                "content-available": True,
-                                "category": "NEW_MESSAGE",
-                                "mutable-content": False,
-                            }
-                        }
-                    }
-                },
-                "validate_only": True,
-            }
-        ),
-        (
-            None,
-            MulticastMessage(
-                tokens=["token-1", "token-2"],
-                data={"text": "hello"},
-            ),
-            {
-                "message": {
-                    "tokens": ["token-1", "token-2"],
-                    "data": {"text": "hello"}
-                },
-                "validate_only": True,
-            }
-        ),
-        (
-            APNSConfig(),
-            MulticastMessage(
-                tokens=["token-1", "token-2"],
-                data={"text": "hello"}
-            ),
-            {
-                "message": {
-                    "tokens": ["token-1", "token-2"],
-                    "data": {"text": "hello"}
-                },
-                "validate_only": True,
-            }
-        ),
-        (
-            APNSConfig(
-                payload=APNSPayload(
-                    aps=Aps(
-                        alert="push-text",
-                        badge=5,
-                        sound="default",
-                        content_available=True,
-                        category="NEW_MESSAGE",
-                        mutable_content=False,
-                    )
-                )
-            ),
-            MulticastMessage(tokens=["token-1", "token-2"], apns=APNSConfig(payload=APNSPayload())),
-            {
-                "message": {
-                    "tokens": ["token-1", "token-2"],
                     "apns": {
                         "payload": {
                             "aps": {